+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 
 - Added `#[extendr(invisible)]` attribute to allow functions to return invisibly in R [[#946]](https://github.com/extendr/extendr/pull/946)
+- An ignore field attribute to the macro `derive(IntoList)` called `#[into_list(ignore)]` [[#864]](https://github.com/extendr/extendr/pull/864)
+- Added `TryFrom<Robj> for Vec<bool>`—an `Error::MustNotBeNA` is returned if an NA is present
+- Added `impl From<&Vec<Robj>> for Robj`
 
 ### Changed